Better bigarray primitives
Change the status of bigarray primitives:
- bigarray_load is now a binary primitive
- bigarray_store is now a ternary primitive

Translation from multi-dimension addressing to linear addressing is
done in the primitive translation, so that the bigarray primitives take
a single linear offset.

Adidtionally the unknwon variant for bigarray kind and layout in
flambda2 are removed because theyr are translated into c-calls in
prepare_lambda.

Cmm code generation has been adapted.