Move Code to its own file and move fields from Function_declaration.t
This moves most of the things in `Function_declaration.t` into `Code.t`,
including arity information, the inlining policy, etc. This makes more
sense in general, since these things don't change from one closure to
another, and closures are the things that use function declarations. It
will also be easier to write Flambda module by hand this way, since we'd
like to declare the arities and such in the `let code` binding rather
than having to repeat them at each closure.

The one way in which the properties can differ between closures is that
a closure may represent a tupled form of the function. Therefore we
retain the `is_tupled` flag on the function declaration.