Fexpr: Use prefixes to distinguish symbols, prims
Since Flambda doesn't distinguish variables from symbols by case, Fexpr
can't either if `-dfexpr` is going to produce usable code. This fixes
the situation by adding `$` to the beginning of symbol names and
removing all case distinctions.

Since we're messing with the lexical syntax anyway, now seemed like a
good time to add `%` to the beginning of the name of each primop, since
there are many primops and we don't want collisions.