New Flambda concrete syntax (#228)
* New Flambda concrete syntax

* Make names mandatory; make literals tagged by default

* Fexpr: Fix printing of tagged immediate as block element

* Flambda syntax: Allow empty block as static constant

* Add/expand support for several important primops

Adds things like tag_imm and get_tag that are all over the place in
real code. The syntax can now support everything from a -dfexpr of
`flambdatest/mlexamples/tests0.ml` (except that cases of variables and
symbols need to be tweaked).

* Fexpr: Use prefixes to distinguish symbols, prims

Since Flambda doesn't distinguish variables from symbols by case, Fexpr
can't either if `-dfexpr` is going to produce usable code. This fixes
the situation by adding `$` to the beginning of symbol names and
removing all case distinctions.

Since we're messing with the lexical syntax anyway, now seemed like a
good time to add `%` to the beginning of the name of each primop, since
there are many primops and we don't want collisions.

* Clean up files from old flambda2.0 directory

* Update .depend