Fexpr: Allow (nearly) arbitrary stings as identifiers
Since things like anonymous functions get weird ids with characters
like `[` and `-` in them, we need to support them in fexpr in order for
`-dflexpect` to work. This lets identifiers start and end with backticks
so that any character (except a backtick or newline) can appear. Pretty
printing still uses the bare forms when possible.

Also fixed some formatting issues that become clear when identifiers are
long.