Check that repr of closure match the new one since 4.12
Starting from ocaml 4.12, closure representaiton changed slightly (see
PR#9619 on the upstream github repo). This commit adds checks in
un_cps_closure to ensure that the computed offsets are conform to the
new representation of closures, mainly that all closure slots precede
all env vars slots, so that the start of environment can be defined.