Convert tests3.ml to .flt
This required adding support for imported symbols (in particular,
external symbols) and the special `Never_returns` result continuation,
as well as primitives for array access (`array_get`, `array_set`, and
`array_length`).

Did some refactoring of the term grammar as well; all keywords are
now prefixed with `KWD_` to avoid clashes (e.g., is a `SYMBOL` token *a*
symbol or the *keyword* `symbol`?).