Sort mutually-recursive constants by linkage name, not id
Since flexpect does a straightforward one-to-one comparison, the order
in which the bindings occur is important even within a recursive group,
and we can't have that order depend on something fragile like what
integer ids things happen to be assigned. Specifically, we get a
spurious test failure from a generated .flt test if the ids in the
original run happen to be different from the ids in the test run.
If we sort them by linkage name instead, the ordering should be
consistent.